ディレクトリ構成をDDD風に変更
